refactor: remove unused components and tests related to shipping guarantee screen
- Deleted GeneralDetails, Instructions, Preview, and Amend components along with their associated HTML, SCSS, and spec files.
- Removed the Create component and its related files, which were previously commented out.
- Cleaned up the Instructions component's HTML and SCSS for better structure and readability.
- Updated the form layout in Instructions to use a two-column grid.
- Ensured all deleted files were properly removed from the project structure.

+++ CC_WEB/src/app/core/services/auth.service.ts
@@ -1,21 +1,36 @@
-import { Injectable } from '@angular/core';
+import { Injectable, PLATFORM_ID, inject } from '@angular/core';
+import { isPlatformBrowser } from '@angular/common';
 
 @Injectable({
   providedIn: 'root'
 })
 export class AuthService {
+  private platformId = inject(PLATFORM_ID);
+
+  private isBrowser(): boolean {
+    return isPlatformBrowser(this.platformId);
+  }
 
   login(userId: string, companyId: string, password: string): boolean {
     // Dummy logic — will replace with API call's later onn
-    if (userId === 'admin' && companyId === 'admin' && password === 'admin') {
-      localStorage.setItem('token', 'dummy-token');
-      localStorage.setItem('role', 'ADMIN');
-      return true;
-    }
+    // if (userId === 'admin' && companyId === 'admin' && password === 'admin') {
+    //   localStorage.setItem('token', 'dummy-token');
+    //   localStorage.setItem('role', 'ADMIN');
+    //   return true;
+    // }
 
-    if (userId === 'user' && companyId === 'user' && password === 'user') {
-      localStorage.setItem('token', 'dummy-token');
-      localStorage.setItem('role', 'USER');
+    // if (userId === 'user' && companyId === 'user' && password === 'user') {
+    //   localStorage.setItem('token', 'dummy-token');
+    //   localStorage.setItem('role', 'USER');
+    //   return true;
+    // }
+        if ((userId === 'admin' && companyId === 'admin' && password === 'admin') ||
+        (userId === 'user' && companyId === 'user' && password === 'user')) {
+
+      if (this.isBrowser()) {
+        localStorage.setItem('token', 'dummy-token');
+        localStorage.setItem('role', userId === 'admin' ? 'ADMIN' : 'USER');
+      }
       return true;
     }
 
@@ -23,14 +38,23 @@ export class AuthService {
   }
 
   logout() {
-    localStorage.clear();
+    if (this.isBrowser()) {
+      localStorage.clear(); 
+    }
   }
 
   checkAuth(): boolean {
+    if (!this.isBrowser()) {
+      return false;
+    }
     return !!localStorage.getItem('token');
   }
 
   getUserCategory(): 'ADMIN' | 'USER' | null {
-    return localStorage.getItem('role') as any;
+    if (!this.isBrowser()) {
+      return null;
+    }
+    // return localStorage.getItem('role') as any;
+    return localStorage.getItem('role') as 'ADMIN' | 'USER' | null;
   }
 }
